fully connected postgress and pgadmin into the project. Add a docker-compose.yaml file to run postgress and pgadmin as containers

+++ Day04-20_JobApp/jobapp/src/main/java/com/example/jobapp/job/impl/JobServiceImpl.java
@@ -5,8 +5,6 @@ import com.example.jobapp.job.JobRepository;
 import com.example.jobapp.job.JobService;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 
@@ -27,6 +25,7 @@ public class JobServiceImpl implements JobService {
     @Override
     public void createJob(Job job) {
         jobRepository.save(job);
+
     }
 
     @Override
@@ -35,33 +34,30 @@ public class JobServiceImpl implements JobService {
     }
 
     @Override
-    public boolean deleteJobById(Long id) {
-        try {
+    public Boolean deleteJobById(Long id) {
+        if (jobRepository.existsById(id)) {
             jobRepository.deleteById(id);
             return true;
-        } catch (Exception e){
+        } else {
             return false;
         }
     }
 
     @Override
-    public boolean updateJobById(Long id, Job updatedJob) {
+    public Boolean updateJob(Long id, Job updatedJob) {
         Optional<Job> jobOptional = jobRepository.findById(id);
 
-
-            if (jobOptional.isPresent()) {
-                Job job = jobOptional.get();
-                job.setTitle(updatedJob.getTitle());
-                job.setDescription(updatedJob.getDescription());
-                job.setMinSalary(updatedJob.getMinSalary());
-                job.setMaxSalary(updatedJob.getMaxSalary());
-                job.setLocation(updatedJob.getLocation());
-
-                jobRepository.save(job);
-                return true;
+        if (jobOptional.isPresent()) {
+            Job job = jobOptional.get();
+            job.setTitle(updatedJob.getTitle());
+            job.setDescription(updatedJob.getDescription());
+            job.setLocation(updatedJob.getLocation());
+            job.setMaxSalary(updatedJob.getMaxSalary());
+            job.setMinSalary(updatedJob.getMinSalary());
+            job.setCompany(updatedJob.getCompany());
+            jobRepository.save(job);
+            return true;
         }
-
         return false;
     }
-
 }
